+++ commands/roleplay/kill.js
@@ -3,11 +3,17 @@ module.exports = {
     usage: "< id / mention >",
     run: async function (client, message, args) {
 
-        if(message.deletable) message.delete();
+        if (client.cooldownManager.checkCooldownAndNotify("rp", message.author.id, message)) {
+            return;
+        }
+
+        if (message.deletable) message.delete();
 
         let killUser = message.mentions.members.first() || message.guild.members.cache.get(args[0]);
         if (!killUser) return message.reply("user not found.");
         if (message.mentions.members.first().user === message.author) return message.reply("you can't roleplay with yourself!");
+
+        client.cooldownManager.setCoolDown("rp", message.author.id, 60);
 
         let array = ["tried to hide the body of", "stabbed and killed", "placed a banana to slip on for", "murdered"];
         await message.channel.send(`${message.author} ${array[Math.round(Math.random() * (array.length - 1))]} ${killUser}!`);
